import data from excel file

+++ src/main/java/com/da/common/ParseDiemTB.java
@@ -1,7 +1,7 @@
 package com.da.common;
 
 public class ParseDiemTB {
-    public static Float diemTB(Float diemmieng, Float diem15p, Float diem90p) {
+    public static Double diemTB(Double diemmieng, Double diem15p, Double diem90p) {
         return ((diemmieng + diem15p) + diem90p * 2) / 4;
     }
 }
